+++ cMake/FreeCAD_Helpers/InitializeFreeCADBuildOptions.cmake
@@ -118,6 +118,7 @@ macro(InitializeFreeCADBuildOptions)
     option(BUILD_BIM "Build the FreeCAD BIM module" ON)
     option(BUILD_DRAFT "Build the FreeCAD draft module" ON)
     option(BUILD_DRAWING "Build the FreeCAD drawing module" OFF)
+    option(BUILD_EXPORT3DPDF "Build the FreeCAD 3D PDF export module" ON)
     option(BUILD_HELP "Build the FreeCAD help module" ON)
     option(BUILD_IDF "Build the FreeCAD idf module" ON)
     option(BUILD_IMPORT "Build the FreeCAD import module" ON)
