+++ src/Mod/CMakeLists.txt
@@ -22,6 +22,10 @@ if(BUILD_DRAFT)
   add_subdirectory(Draft)
 endif(BUILD_DRAFT)
 
+if(BUILD_EXPORT3DPDF)
+  add_subdirectory(Export3DPDF)
+endif(BUILD_EXPORT3DPDF)
+
 if(BUILD_FEM)
     add_subdirectory(Fem)
 endif(BUILD_FEM)
