+++ src/Mod/Export3DPDF/App/AppExport3DPDF.cpp
@@ -1,0 +1,57 @@
+/***************************************************************************
+ *   Copyright (c) 2024 FreeCAD Developers                                 *
+ *                                                                         *
+ *   This file is part of the FreeCAD CAx development system.              *
+ *                                                                         *
+ *   This library is free software; you can redistribute it and/or         *
+ *   modify it under the terms of the GNU Library General Public           *
+ *   License as published by the Free Software Foundation; either          *
+ *   version 2 of the License, or (at your option) any later version.      *
+ *                                                                         *
+ *   This library  is distributed in the hope that it will be useful,      *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU Library General Public License for more details.                  *
+ *                                                                         *
+ *   You should have received a copy of the GNU Library General Public     *
+ *   License along with this library; see the file COPYING.LIB. If not,    *
+ *   write to the Free Software Foundation, Inc., 59 Temple Place,         *
+ *   Suite 330, Boston, MA  02111-1307, USA                                *
+ *                                                                         *
+ ***************************************************************************/
+
+#include "PreCompiled.h"
+
+#include <CXX/Extensions.hxx>
+#include <Base/Console.h>
+#include <Base/Interpreter.h>
+#include <Base/PyObjectBase.h>
+
+#include "../Export3DPDFGlobal.h"
+
+namespace Export3DPDF 
+{
+
+class Module : public Py::ExtensionModule<Module>
+{
+public:
+    Module() : Py::ExtensionModule<Module>("Export3DPDF")
+    {
+        initialize("This module is the Export3DPDF module."); // register with Python
+    }
+
+private:
+};
+
+extern PyObject* initModule();
+
+} // namespace Export3DPDF
+
+/* Python entry */
+PyMOD_INIT_FUNC(Export3DPDF)
+{
+    PyObject* mod = Export3DPDF::initModule();
+    Base::Console().log("Loading Export3DPDF module... done\n");
+
+    PyMOD_Return(mod);
+}
