+++ src/Mod/Export3DPDF/Gui/AppExport3DPDFGui.cpp
@@ -1,0 +1,72 @@
+/***************************************************************************
+ *   Copyright (c) 2024 FreeCAD Developers                                 *
+ *                                                                         *
+ *   This file is part of the FreeCAD CAx development system.              *
+ *                                                                         *
+ *   This library is free software; you can redistribute it and/or         *
+ *   modify it under the terms of the GNU Library General Public           *
+ *   License as published by the Free Software Foundation; either          *
+ *   version 2 of the License, or (at your option) any later version.      *
+ *                                                                         *
+ *   This library  is distributed in the hope that it will be useful,      *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU Library General Public License for more details.                  *
+ *                                                                         *
+ *   You should have received a copy of the GNU Library General Public     *
+ *   License along with this library; see the file COPYING.LIB. If not,    *
+ *   write to the Free Software Foundation, Inc., 59 Temple Place,         *
+ *   Suite 330, Boston, MA  02111-1307, USA                                *
+ *                                                                         *
+ ***************************************************************************/
+
+#include "PreCompiled.h"
+
+#include <CXX/Extensions.hxx>
+#include <Base/Console.h>
+#include <Base/Interpreter.h>
+#include <Base/PyObjectBase.h>
+#include <Gui/Application.h>
+
+#include "../Export3DPDFGlobal.h"
+#include "Command.h"
+
+// Commands for 3D PDF Export
+void CreateExport3DPDFCommands();
+
+namespace Export3DPDFGui 
+{
+extern PyObject* initModule();
+}
+
+/* Python entry */
+PyMOD_INIT_FUNC(Export3DPDFGui)
+{
+    if (!Gui::Application::Instance) {
+        PyErr_SetString(PyExc_ImportError, "Cannot load Gui module in console application.");
+        PyMOD_Return(nullptr);
+    }
+
+    // load dependent module
+    try {
+        Base::Interpreter().loadModule("Export3DPDF");
+    }
+    catch (const Base::Exception& e) {
+        PyErr_SetString(PyExc_ImportError, e.what());
+        PyMOD_Return(nullptr);
+    }
+    
+    PyObject* mod = Export3DPDFGui::initModule();
+    Base::Console().log("Loading GUI of Export3DPDF module... done\n");
+
+    // instantiating the commands
+    CreateExport3DPDFCommands();
+
+    PyMOD_Return(mod);
+}
+
+void CreateExport3DPDFCommands()
+{
+    Gui::CommandManager &rcCmdMgr = Gui::Application::Instance->commandManager();
+    rcCmdMgr.addCommand(new StdCmdPrint3dPdf());
+}
