+++ src/Mod/Export3DPDF/Gui/CMakeLists.txt
@@ -1,0 +1,43 @@
+include_directories(
+    ${CMAKE_BINARY_DIR}
+    ${CMAKE_SOURCE_DIR}/src
+    ${CMAKE_BINARY_DIR}/src
+    ${CMAKE_CURRENT_BINARY_DIR}
+)
+
+set(Export3DPDFGui_LIBS
+    FreeCADGui
+    Export3DPDF
+)
+
+SET(Export3DPDFGui_SRCS
+    AppExport3DPDFGui.cpp
+    AppExport3DPDFGuiPy.cpp
+    Command.cpp
+    Command.h
+    Export3DPDFGui.cpp
+    Export3DPDFGui.h
+    PreCompiled.cpp
+    PreCompiled.h
+)
+
+if(FREECAD_USE_PCH)
+    add_definitions(-D_PreComp_)
+    GET_MSVC_PRECOMPILED_SOURCE("PreCompiled.cpp" PCH_SRCS ${Export3DPDFGui_SRCS})
+    ADD_MSVC_PRECOMPILED_HEADER(Export3DPDFGui PreCompiled.h PreCompiled.cpp PCH_SRCS)
+endif(FREECAD_USE_PCH)
+
+add_library(Export3DPDFGui SHARED ${Export3DPDFGui_SRCS})
+target_link_libraries(Export3DPDFGui ${Export3DPDFGui_LIBS})
+
+# Ensure FreeCADGui is built before Export3DPDFGui
+add_dependencies(Export3DPDFGui FreeCADGui)
+
+if (FREECAD_WARN_ERROR)
+    target_compile_warn_error(Export3DPDFGui)
+endif()
+
+SET_BIN_DIR(Export3DPDFGui Export3DPDFGui /Mod/Export3DPDF)
+SET_PYTHON_PREFIX_SUFFIX(Export3DPDFGui)
+
+INSTALL(TARGETS Export3DPDFGui DESTINATION ${CMAKE_INSTALL_LIBDIR})
