+++ src/Mod/Export3DPDF/Gui/Export3DPDFGui.cpp
@@ -1,0 +1,38 @@
+/***************************************************************************
+ *   Copyright (c) 2024 FreeCAD Developers                                 *
+ *                                                                         *
+ *   This file is part of the FreeCAD CAx development system.              *
+ *                                                                         *
+ *   This library is free software; you can redistribute it and/or         *
+ *   modify it under the terms of the GNU Library General Public           *
+ *   License as published by the Free Software Foundation; either          *
+ *   version 2 of the License, or (at your option) any later version.      *
+ *                                                                         *
+ *   This library  is distributed in the hope that it will be useful,      *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU Library General Public License for more details.                  *
+ *                                                                         *
+ *   You should have received a copy of the GNU Library General Public     *
+ *   License along with this library; see the file COPYING.LIB. If not,    *
+ *   write to the Free Software Foundation, Inc., 59 Temple Place,         *
+ *   Suite 330, Boston, MA  02111-1307, USA                                *
+ *                                                                         *
+ ***************************************************************************/
+
+#include "PreCompiled.h"
+
+#include <Base/Console.h>
+#include "Export3DPDFGui.h"
+
+using namespace Export3DPDFGui;
+
+void Export3DPDFGuiUtils::initialize()
+{
+    Base::Console().log("Initializing Export3DPDF GUI utilities...\n");
+    
+    // Future GUI initialization code can go here
+    // For example: setting up icons, preferences, etc.
+    
+    Base::Console().log("Export3DPDF GUI utilities initialized.\n");
+}
